requirements.txt
flask>=3.0
flask-cors>=4.0
moviepy>=1.0.3
opencv-python-headless>=4.8
scenedetect[opencv]>=0.6
openai-whisper>=20231117
anthropic>=0.40
numpy>=1.24
Pillow>=10.0
